# perf: improve user specified flush delay automatically by statistic