@rueian 2022-12-14
related to https://github.com/rueian/rueidis/issues/156