@FZambia 2022-12-14
Hello @rueian  - this looks interesting, some initial thoughts after looking at PR (have not even run it yet):

* Is it possible to describe the logic behind this formula?
* To be fair I am a bit nervous at the moment to enable non-deterministic connection behaviour with auto adjustment – `reasonable` CPU is a bit subjective matter, so I think automatic adjustment should be optional