@FZambia 2022-12-14
Another though: let's say we have a bigger latency, and set some delay we are comfortable with in the app. As far as I understand this auto adjustment does not take latency into account - possible there is no much sense to reduce delay automatically given large RTT time (since it may result in larger CPU on Redis at the end).

BTW, I experimented locally with https://github.com/Shopify/toxiproxy yesterday to add some latency to Redis. But I am on MacOS, for Linux `tc` may be more obvious choice - https://daniel.haxx.se/blog/2010/12/14/add-latency-to-localhost/.