@rueian 2022-12-15
> `reasonable` CPU is a bit subjective matte

Indeed. I use the word `reasonable` because the CPU usage of this approach is similar to redigo's.

> Is it possible to describe the logic behind this formula?

Let's start with the question:

_**How long should we wait for more commands?**_

Sure, we could just sleep for `MaxFlushDelay` specified by the user and have the lowest CPU usage. But it could easily wait too long and result in very poor throughput.

For example, sleeping for 100 microseconds reduced the throughput by ~290%.

```sh
▶ benchstat ori-0us-p1.txt old-100us-p1.txt
name        old time/op    new time/op    delta
Rueidis-10    3.08µs ± 0%   11.99µs ± 0%  +289.49%  (p=0.000 n=19+20)
```

To avoid sleeping too long, I would like to make this sleep including the successive flush similar to previous flushes. In other words, I want `time(sleep) + time(flush) == avg time(previous flushes)`.

Assuming the `time(flush)` is proportional to how many bytes it flushes, then we can have average flush time of a byte.