@rueian 2022-12-15
> But I suppose if I wanted to improve throughput - I could just tune the MaxFlushDelay parameter myself as I know my app and its load as a developer better.

Sure, but there are many cases that their traffic pattern can't be fit to a fixed `MaxFlushDelay`. Adaptive approach may still be good to have.

> The point of adding delay in my case was to reduce CPU, but still have acceptable throughput. 

We probably can also have a mode allowing user to set `acceptable throughput` then trying to reduce CPU usage.

> Maybe I am too conservative in that – but just prefer keeping things as simple as possible.

Don't worry. This won't be merged until we find a better solution.