@FZambia 2022-12-16
Current `MaxFlushDelay`, while may be a bit `manual`, provides a working understandable way to tradeoff throughput/CPU. And it works deterministically. Maybe value may be not very good for some setup – but possible to tune. I'd say – it's already like `acceptable latency` (well, not exactly of course, more like `acceptable delay` but it seems good enough to solve a problem).

I think you implemented some bright ideas here, but I still not convinced that I want to use this mode somehow. Probably it's better to wait for more feedback here, maybe from other users.