@rueian 2022-12-17
> Thinking more.. Not sure whether it's good idea or not – can this batching strategy be a pluggable element somehow – just to keep Rueidis core simple yet extensible

It is a good idea. Probably something like this `func(ctx CurrentFlushContext) time.Duration` allows users to choose how long they want to wait based on some information.